requirements.txt
flask==2.3.3
groq
python-dotenv==1.0.0
httpx==0.24.1

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Linting
flake8==6.1.0
black==23.7.0

# Type checking
mypy==1.5.0
types-requests==2.31.0.1

# Security
bandit==1.7.5